requirements.txt
casadi==3.6.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
imageio==2.34.2
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy-loader==0.4
markupsafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opengen==0.8.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
py==1.11.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.2
retry==0.9.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.2.0
shapely==2.0.5
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.10
torch==2.4.0
torchvision==0.19.0
triton==3.0.0
typing-extensions==4.12.2
tzdata==2024.1
